+++ clients/admin-ui/src/features/common/form/inputs.tsx
@@ -68,6 +68,7 @@ export interface CustomInputProps {
   isRequired?: boolean;
   textColor?: string;
   inputRightElement?: React.ReactNode;
+  size?: string;
 }
 
 // We allow `undefined` here and leave it up to each component that uses this field
@@ -95,6 +96,7 @@ export const TextInput = forwardRef(
     {
       isPassword,
       inputRightElement,
+      size,
       ...props
     }: InputProps & {
       isPassword: boolean;
@@ -110,7 +112,7 @@ export const TextInput = forwardRef(
       setType(type === "password" ? "text" : "password");
 
     return (
-      <InputGroup size="sm">
+      <InputGroup size={size ?? "sm"}>
         <Input
           {...props}
           ref={ref as LegacyRef<HTMLInputElement> | undefined}
@@ -529,6 +531,7 @@ export const CustomTextInput = ({
   variant = "inline",
   isRequired = false,
   inputRightElement,
+  size,
   ...props
 }: CustomInputProps & StringField) => {
   const [initialField, meta] = useField(props);
@@ -546,6 +549,7 @@ export const CustomTextInput = ({
       placeholder={placeholder}
       isPassword={isPassword}
       inputRightElement={inputRightElement}
+      size={size}
     />
   );
 
@@ -576,7 +580,12 @@ export const CustomTextInput = ({
       <VStack alignItems="start">
         {label ? (
           <Flex alignItems="center">
-            <Label htmlFor={props.id || props.name} fontSize="xs" my={0} mr={1}>
+            <Label
+              htmlFor={props.id || props.name}
+              fontSize={size ?? "xs"}
+              my={0}
+              mr={1}
+            >
               {label}
             </Label>
             {tooltip ? <QuestionTooltip label={tooltip} /> : null}