+++ clients/admin-ui/src/features/user-management/NewUserForm.tsx
@@ -4,6 +4,7 @@ import React, { useEffect } from "react";
 import UserManagementTabs from "user-management/UserManagementTabs";
 
 import { useAppDispatch } from "~/app/hooks";
+import { isErrorResult } from "~/features/common/helpers";
 import { USER_MANAGEMENT_ROUTE } from "~/features/common/nav/v2/routes";
 
 import {
@@ -22,13 +23,19 @@ const NewUserForm = () => {
   }, [dispatch]);
 
   const handleSubmit = async (values: FormValues) => {
-    const b64Password = utf8ToB64(values.password);
+    const b64Password = values.password
+      ? utf8ToB64(values.password)
+      : undefined;
     const result = await createUser({
       ...values,
       password: b64Password,
-    }).unwrap();
-    router.push(`${USER_MANAGEMENT_ROUTE}/profile/${result.id}`);
+    });
+    if (!isErrorResult(result)) {
+      router.push(`${USER_MANAGEMENT_ROUTE}/profile/${result.data.id}`);
+    }
+    return result;
   };
+
   return (
     <div>
       <main>