+++ src/fides/api/schemas/user.py
@@ -1,8 +1,9 @@
 import re
 from datetime import datetime
+from enum import Enum
 from typing import Optional
 
-from pydantic import validator
+from pydantic import EmailStr, validator
 
 from fides.api.cryptography.cryptographic_util import decode_password
 from fides.api.schemas.base_class import FidesSchema
@@ -20,14 +21,16 @@ class UserCreate(FidesSchema):
     """Data required to create a FidesUser."""
 
     username: str
-    password: str
+    password: Optional[str]
+    email_address: EmailStr
     first_name: Optional[str]
     last_name: Optional[str]
+    disabled: bool = False
 
     @validator("username")
     @classmethod
     def validate_username(cls, username: str) -> str:
-        """Ensure password does not have spaces."""
+        """Ensure username does not have spaces."""
         if " " in username:
             raise ValueError("Usernames cannot have spaces.")
         return username
@@ -79,8 +82,11 @@ class UserResponse(FidesSchema):
     id: str
     username: str
     created_at: datetime
+    email_address: Optional[EmailStr]
     first_name: Optional[str]
     last_name: Optional[str]
+    disabled: Optional[bool] = False
+    disabled_reason: Optional[str]
 
 
 class UserLoginResponse(FidesSchema):
@@ -104,7 +110,14 @@ class UserForcePasswordReset(FidesSchema):
 
 
 class UserUpdate(FidesSchema):
-    """Data required to update a FidesopsUser"""
+    """Data required to update a FidesUser"""
 
+    email_address: Optional[EmailStr]
     first_name: Optional[str]
     last_name: Optional[str]
+
+
+class DisabledReason(Enum):
+    """Reasons for why a user is disabled"""
+
+    pending_invite = "pending_invite"