+++ src/fides/api/service/messaging/message_dispatch_service.py
@@ -39,6 +39,7 @@ from fides.api.schemas.messaging.messaging import (
     RequestReceiptBodyParams,
     RequestReviewDenyBodyParams,
     SubjectIdentityVerificationBodyParams,
+    UserInviteBodyParams,
 )
 from fides.api.schemas.redis_cache import Identity
 from fides.api.service.messaging.messaging_crud_service import (
@@ -128,6 +129,7 @@ def dispatch_message(
             RequestReceiptBodyParams,
             RequestReviewDenyBodyParams,
             ErasureRequestBodyParams,
+            UserInviteBodyParams,
         ]
     ] = None,
     subject_override: Optional[str] = None,
@@ -155,8 +157,11 @@ def dispatch_message(
     logger.info("Getting custom messaging template for action type: {}", action_type)
     messaging_template = get_messaging_template_by_key(db=db, key=action_type.value)
 
+    config_proxy = ConfigProxy(db=db)
+
     if messaging_method == MessagingMethod.EMAIL:
         message = _build_email(
+            config_proxy=config_proxy,
             action_type=action_type,
             body_params=message_body_params,
             messaging_template=messaging_template,
@@ -267,6 +272,7 @@ def _render(template_str: str, variables: Optional[Dict] = None) -> str:
 
 
 def _build_email(  # pylint: disable=too-many-return-statements
+    config_proxy: ConfigProxy,
     action_type: MessagingActionType,
     body_params: Any,
     messaging_template: Optional[MessagingTemplate] = None,
@@ -360,6 +366,18 @@ def _build_email(  # pylint: disable=too-many-return-statements
         base_template = get_email_template(action_type)
         return EmailForActionType(
             subject="Test message from fides", body=base_template.render()
+        )
+    if action_type == MessagingActionType.USER_INVITE:
+        base_template = get_email_template(action_type)
+        return EmailForActionType(
+            subject="Welcome to Fides",
+            body=base_template.render(
+                {
+                    "admin_ui_url": config_proxy.admin_ui.url,
+                    "username": body_params.username,
+                    "invite_code": body_params.invite_code,
+                }
+            ),
         )
     logger.error("Message action type {} is not implemented", action_type)
     raise MessageDispatchException(