+++ src/fides/config/admin_ui_settings.py
@@ -1,4 +1,6 @@
-from pydantic import Field
+from typing import Optional
+
+from pydantic import AnyHttpUrl, Field
 
 from .fides_settings import FidesSettings
 
@@ -8,6 +10,9 @@ class AdminUISettings(FidesSettings):
 
     enabled: bool = Field(
         default=True, description="Toggle whether the Admin UI is served."
+    )
+    url: Optional[AnyHttpUrl] = Field(
+        default=None, description="The base URL for the Admin UI."
     )
 
     class Config: