+++ tests/conftest.py
@@ -266,6 +266,7 @@ def application_user(db, oauth_client):
         data={
             "username": unique_username,
             "password": "test_password",
+            "email_address": f"{unique_username}@ethyca.com",
             "first_name": "Test",
             "last_name": "User",
         },
@@ -282,6 +283,7 @@ def user(db):
         data={
             "username": "test_fidesops_user",
             "password": "TESTdcnG@wzJeu0&%3Qe2fGo7",
+            "email_address": "fides.user@ethyca.com",
         },
     )
     client = ClientDetail(
@@ -904,6 +906,7 @@ def owner_user(db):
         data={
             "username": "test_fides_owner_user",
             "password": "TESTdcnG@wzJeu0&%3Qe2fGo7",
+            "email_address": "owner.user@ethyca.com",
         },
     )
     client = ClientDetail(
@@ -930,6 +933,7 @@ def approver_user(db):
         data={
             "username": "test_fides_viewer_user",
             "password": "TESTdcnG@wzJeu0&%3Qe2fGo7",
+            "email_address": "approver.user@ethyca.com",
         },
     )
     client = ClientDetail(
@@ -956,6 +960,7 @@ def viewer_user(db):
         data={
             "username": "test_fides_viewer_user",
             "password": "TESTdcnG@wzJeu0&%3Qe2fGo7",
+            "email_address": "viewer2.user@ethyca.com",
         },
     )
     client = ClientDetail(
@@ -981,6 +986,7 @@ def contributor_user(db):
         data={
             "username": "test_fides_contributor_user",
             "password": "TESTdcnG@wzJeu0&%3Qe2fGo7",
+            "email_address": "contributor.user@ethyca.com",
         },
     )
     client = ClientDetail(
@@ -1009,6 +1015,7 @@ def viewer_and_approver_user(db):
         data={
             "username": "test_fides_viewer_and_approver_user",
             "password": "TESTdcnG@wzJeu0&%3Qe2fGo7",
+            "email_address": "viewerapprover.user@ethyca.com",
         },
     )
     client = ClientDetail(