+++ tests/ops/service/messaging/message_dispatch_service_test.py
@@ -25,6 +25,7 @@ from fides.api.schemas.messaging.messaging import (
     MessagingServiceType,
     RequestReviewDenyBodyParams,
     SubjectIdentityVerificationBodyParams,
+    UserInviteBodyParams,
 )
 from fides.api.schemas.privacy_notice import PrivacyNoticeHistorySchema
 from fides.api.schemas.privacy_preference import MinimalPrivacyPreferenceHistorySchema
@@ -39,6 +40,7 @@ from fides.api.service.messaging.message_dispatch_service import (
     _twilio_sms_dispatcher,
     dispatch_message,
 )
+from fides.config import CONFIG
 
 
 @pytest.fixture
@@ -582,6 +584,44 @@ class TestMessageDispatchService:
                 body=body,
             ),
             "sovrn_test@example.com",
+        )
+
+    @pytest.fixture
+    def mock_config_admin_ui_url(self):
+        admin_ui_url = CONFIG.admin_ui.url
+        CONFIG.admin_ui.url = "http://localhost:3000"
+        yield
+        CONFIG.admin_ui.url = admin_ui_url
+
+    @mock.patch(
+        "fides.api.service.messaging.message_dispatch_service._mailgun_dispatcher"
+    )
+    def test_email_dispatch_user_invite_email(
+        self,
+        mock_mailgun_dispatcher: Mock,
+        db: Session,
+        messaging_config,
+        mock_config_admin_ui_url,
+    ) -> None:
+        dispatch_message(
+            db=db,
+            action_type=MessagingActionType.USER_INVITE,
+            to_identity=Identity(**{"email": "test@example.com"}),
+            service_type=MessagingServiceType.mailgun.value,
+            message_body_params=UserInviteBodyParams(
+                username="test", invite_code="123"
+            ),
+        )
+
+        body = '<!DOCTYPE html>\n<html lang="en">\n  <head>\n    <meta charset="UTF-8" />\n    <title>Welcome to Fides</title>\n  </head>\n  <body>\n    <main>\n      <p>You\'ve been invited to join Fides, click <a href=http://localhost:3000/login?invite_code=123&username=test>here</a> to accept the invite and setup your account.</p>\n    </main>\n  </body>\n</html>'
+
+        mock_mailgun_dispatcher.assert_called_with(
+            messaging_config,
+            EmailForActionType(
+                subject="Welcome to Fides",
+                body=body,
+            ),
+            "test@example.com",
         )